requirements.txt
numpy
openai-whisper
ollama
sounddevice
soundfile
asyncio
langchain
kokoro
ipython
